Separated KeyValueMaster into an interface and an implementation
