+++ keyvalue/server/src/main/java/alluxio/master/keyvalue/KeyValueMasterFactory.java
@@ -49,13 +49,13 @@ public final class KeyValueMasterFactory implements MasterFactory {
   }
 
   @Override
-  public KeyValueMaster create(MasterRegistry registry, JournalFactory journalFactory) {
+  public DefaultKeyValueMaster create(MasterRegistry registry, JournalFactory journalFactory) {
     Preconditions.checkArgument(journalFactory != null, "journal factory may not be null");
     LOG.info("Creating {} ", KeyValueMaster.class.getName());
     Journal journal = journalFactory.create(getName());
     FileSystemMaster fileSystemMaster = registry.get(FileSystemMaster.class);
-    KeyValueMaster keyValueMaster = new KeyValueMaster(fileSystemMaster, journal);
-    registry.add(KeyValueMaster.class, keyValueMaster);
-    return keyValueMaster;
+    DefaultKeyValueMaster defaultKeyValueMaster = new DefaultKeyValueMaster(fileSystemMaster, journal);
+    registry.add(DefaultKeyValueMaster.class, defaultKeyValueMaster);
+    return defaultKeyValueMaster;
   }
 }
